mama-synth/mama-synth-gc-eval/requirements.txt
numpy>=1.24
SimpleITK>=2.2
scikit-learn>=1.2
scipy>=1.10
scikit-image>=0.20
pyradiomics>=3.1
lpips>=0.1
